examples/apps/snowflake-integration/requirements.txt
streamlit
Pillow~=9.3.0
aiohttp>=3.7.4
pandas~=1.3.4

# BEGIN: Snowflake requirements file
# Generated on: Python 3.9.15
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
filelock==3.8.0
idna==3.4
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.6.0
pyOpenSSL==23.1.0
pytz==2022.6
requests==2.31.0
typing_extensions==4.4.0
urllib3==1.26.12
snowflake-connector-python==3.0.2
# END: Snowflake requirements file
########################## SNOWFLAKE End ##########################

